+++ core-default-node/templates/_Gruntfile.js
@@ -25,6 +25,7 @@ Other calls (relatively in order of importantance / most used). Scroll to the bo
 - dev
 	`grunt dev` for watching and auto-running build and tests
 	`grunt dev-test` for watching and auto-running TESTS only
+	`grunt dev-karma-cov` for watching and auto-running karma unit test COVERAGE (since dev-test does NOT run coverage due to issue..). To see / generate coverage, run this AND 'dev-test' OR 'dev' (in 2 separate command windows)
 	`grunt dev-build` for watching and auto-running BUILD (i.e. `grunt q`) only
 - test
 	`grunt karma-cov` to run/build karma/angular coverage report (since grunt dev watch task does NOT do this due to a bug/issue with karma where running the coverage does NOT show test info on the console, which makes it annoying to debug)
@@ -322,8 +323,8 @@ module.exports = function(grunt) {
 						prefix: publicPathRelativeDot,
 						moduleGroup: 'allNoBuildCss',
 						outputFiles: {
-							js: ['watch.karmaUnitJs.files', 'watch.karmaUnit.files'],
-							testUnit: ['watch.karmaUnitTest.files', 'watch.karmaUnit.files']
+							js: ['watch.karmaUnitJs.files', 'watch.karmaUnit.files', 'watch.karmaCovUnit.files'],
+							testUnit: ['watch.karmaUnitTest.files', 'watch.karmaUnit.files', 'watch.karmaCovUnit.files']
 						}
 					},
 					//index.html file paths (have the static path prefix for use in <link rel="stylesheet" > and <script> tags)
@@ -754,6 +755,9 @@ module.exports = function(grunt) {
 				test: {
 					include: ['karmaUnit']
 				},
+				testKarmaCov: {
+					include: ['karmaCovUnit']
+				},
 				all: {
 					include: ['build', 'karmaUnit']
 				}
@@ -775,6 +779,10 @@ module.exports = function(grunt) {
 				karmaUnit: {
 					files: [],		//will be filled by grunt-buildfiles
 					tasks: ['karma:watch:run']
+				},
+				karmaCovUnit: {
+					files: [],		//will be filled by grunt-buildfiles
+					tasks: ['karma-cov']
 				},
 				
 				//run buildfiles pretty much any time a file changes (since this generates file lists for other tasks - will this work / update them since grunt is already running??)
@@ -1022,7 +1030,7 @@ module.exports = function(grunt) {
 		//shorthand for 'shell:protractor' (this assumes node & selenium servers are already running)
 		grunt.registerTask('e2e', ['shell:protractor']);
 		
-		grunt.registerTask('karma-cov', ['karma:unit', 'coverage']);
+		grunt.registerTask('karma-cov', ['clean', 'karma:unit', 'coverage']);
 		
 		grunt.registerTask('test-frontend', ['karma-cov', 'e2e']);
 
@@ -1155,6 +1163,9 @@ module.exports = function(grunt) {
 		//test only
 		// grunt.registerTask('dev-test', ['q-watch', 'test-cleanup', 'test-setup', 'karma:watch:start', 'watch:karmaUnitJs', 'watch:karmaUnitTest']);		//doesn't work - watch isn't a multi-task..
 		grunt.registerTask('dev-test', ['q-watch', 'test-cleanup', 'test-setup', 'karma:watch:start', 'focus:test']);
+		
+		// grunt.registerTask('dev-karma-cov', ['q-watch', 'karma:watchCov:start', 'focus:testKarmaCov']);		//does not work.. get 'app/src/coverage-angular' does not exist..
+		grunt.registerTask('dev-karma-cov', ['q-watch', 'focus:testKarmaCov']);
 		
 		//build only
 		// grunt.registerTask('dev-build', ['q-watch', 'watch:buildfiles', 'watch:html', 'watch:less', 'watch:jsHintFrontend', 'watch:jsHintBackend']);		//doesn't work - watch isn't a multi-task..		//@todo - if do change this, make sure to add scss version!!